+++ pillow_heif/options.py
@@ -18,7 +18,7 @@ When use pillow_heif as a plugin you can set it with: `register_*_opener(thumbna
 DEPTH_IMAGES = True
 """Option to enable/disable depth image support
 
-When use pillow_heif as a plugin you can set it with: `register_*_opener(depths=False)`"""
+When use pillow_heif as a plugin you can set it with: `register_*_opener(depth_images=False)`"""
 
 
 QUALITY = None
@@ -48,3 +48,12 @@ ALLOW_INCORRECT_HEADERS = False
 To learn more read: `here <https://github.com/strukturag/libheif/issues/784>`_
 
 When use pillow_heif as a plugin you can set it with: `register_*_opener(allow_incorrect_headers=True)`"""
+
+
+SAVE_NCLX_PROFILE = False
+"""Should be `nclx` profile saved or not.
+
+.. note:: macOS viewer ``does not support opening`` images with **nclx** profiles, so this is disabled by default.
+.. note:: `save_nclx_profile` specified during calling ``save`` has higher priority then this.
+
+When use pillow_heif as a plugin you can set it with: `register_*_opener(save_nclx_profile=True)`"""